final/backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.20.32
botocore==1.23.32
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.21.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.3.2
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==9.0.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
